pyproject.toml
[project]
name = "katasym"
version = "0.1.0"
description = "Python client for the Katasymbol E10 BLE thermal label printer (reverse-engineered)."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Giorgio Campiotti" }]
keywords = ["bluetooth", "ble", "thermal-printer", "katasymbol", "e10", "label"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Topic :: Printing",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "bleak>=0.22",
    "pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[project.scripts]
katasym = "katasym.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["katasym"]

[tool.hatch.build.targets.sdist]
include = ["katasym", "examples", "README.md"]

[tool.ruff]
line-length = 100
target-version = "py310"
